rag/chunking/llamaindex/requirements.txt
llama-index-core
beautifulsoup4
